Rip out unified queries file and co-locate code
