+++ packages/frontend-airdos/src/pages/GroupPage/GroupPageQueries.tsx
@@ -1,0 +1,15 @@
+import gql from "graphql-tag"
+
+export const GET_GROUP_BY_NAME = gql`
+  query getGroupByName($name: String!) {
+    groupByName(name: $name) {
+      name
+      displayName
+      posts {
+        location
+        text
+      }
+      members
+    }
+  }
+`
